[FEATURE] Added another docker-compose.yml to run necessary services in Docker and Spring services on localhost
Added a new docker-compose.yml file to manage the necessary services for the project. This configuration allows the essential services to run in Docker while enabling the Spring services to operate on localhost. This setup supports a development environment where the backend services can be tested and debugged locally while other components are managed in Docker.